+++ portal/autoconfig.py
@@ -133,6 +133,7 @@ SETTINGS = {
         "django.contrib.auth.backends.ModelBackend",
         "portal.backends.StudentLoginBackend",
     ],
+    "USE_TZ": True,
     # ----------------------------------------------------------------------------------
     # CSP CONFIG
     # ----------------------------------------------------------------------------------